+++ CHANGELOG.rst
@@ -42,6 +42,8 @@ Fixed
 
 - **General**
     - ``LegacyKeyValueFormat`` warnings in Docker build (#2089)
+- **Landingzones**
+    - Zone delete timeline status not updated with missing collection (#2096)
 
 Removed
 -------